create oods module, remove circle m31 as it is never used
