+++ src/circle/mod.rs
@@ -1,2 +1,4 @@
 mod bitcoin_script;
 pub use bitcoin_script::*;
+
+// TODO(Bitcoin): implement the non-bitcoin-script version of the OODS sampling.
