--- src/circle_secure/mod.rs
@@ -1,4 +1,0 @@
-mod bitcoin_script;
-pub use bitcoin_script::*;
-
-// TODO(Bitcoin): implement the non-bitcoin-script version of the OODS sampling.
